updated style

+++ controllers/postcontroller.js
@@ -107,7 +107,7 @@ module.exports.destroy= async function(req,res)
                // CHANGE :: delete the associated likes for the post and all its comments' likes too
                 await Like.deleteMany({likeable: post, onModel: 'Post'});
                 await Like.deleteMany({_id: {$in: post.comment}});
-                fs.unlinkSync(path.join(__dirname,'..',post.Postfile));
+               if(post.Postfile){ fs.unlinkSync(path.join(__dirname,'..',post.Postfile));}
                 post.remove();
                 await  Comment.deleteMany({post:req.params.id});
